piplib/acs-zeep-client/requirements.txt
httpx>=0.25.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
typer>=0.12.0,<1.0.0
rich>=13.0.0,<15.0.0
